tests/gemini_test/pyproject.toml
[project]
name = "gemini-live-simple-voice-agent-test"
version = "0.1.0"
description = "Minimal ADK Gemini Live API voice+vision+text test app"
readme = "adk-samples/README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "google-adk>=1.20.0",
    "google-genai>=1.36.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
